+++ co-log/CHANGELOG.md
@@ -3,6 +3,12 @@
 `co-log` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
+## Unreleased
+
+* [#144](https://github.com/kowainik/co-log/issues/144):
+  Add Windows CI check.
+* Use `chronos-1.1` as `1.0.9` is not Windows-competitive.
+
 ## 0.3.0.0 — May 5, 2019
 
 * [#77](https://github.com/kowainik/co-log/issues/77):