+++ co-log/CHANGELOG.md
@@ -3,16 +3,37 @@
 `co-log` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
-## Unreleased
+## 0.4.0.0 — Jan 19, 2020
 
 * [#120](https://github.com/kowainik/co-log/issues/120):
-  Improve time formatting. Instead of `29-12-2019 22:00:00.000`
-  it looks like `29 Dec 2019 22:00:00.000 +00:00` now.
+  Improve time formatting.
+
+  Old: `29-12-2019 22:00:00.000`
+
+  New: `29 Dec 2019 22:00:00.000 +00:00`
+
+  (by [@vrom911](https://github.com/vrom911))
 * [#144](https://github.com/kowainik/co-log/issues/144):
   Add Windows CI check.
+  (by [@vrom911](https://github.com/vrom911))
+* [#148](https://github.com/kowainik/co-log/issues/148):
+  Support GHC-8.8.2.
+  (by [@chshersh](https://github.com/chshersh))
+* [#119](https://github.com/kowainik/co-log/issues/119):
+  Add new message type that allows printing messages without
+  `Severity`.
+  (by [@sphaso](https://github.com/sphaso))
+* Use `chronos-1.1` as `1.0.9` is not Windows-compatible.
+  (by [@vrom911](https://github.com/vrom911))
 * [#156](https://github.com/kowainik/co-log/issues/156):
   Improve documentation for the `Colog.Concurrent` module.
-* Use `chronos-1.1` as `1.0.9` is not Windows-competitive.
+  (by [@chshersh](https://github.com/chshersh))
+* [#146](https://github.com/kowainik/co-log/issues/146):
+  Allow `ansi-terminal-0.10`.
+  (by [@mpilgrem](https://github.com/mpilgrem))
+* [#124](https://github.com/kowainik/co-log/issues/124):
+  Implement executable playground for concurrent logging.
+  (by [@chshersh](https://github.com/chshersh))
 
 ## 0.3.0.0 — May 5, 2019